requirements.txt
asgiref==3.8.1
coverage==7.6.10
Django==5.1.4
django-rest-framework==0.1.0
djangorestframework==3.15.2
flake8==7.1.1
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
python-dotenv==1.0.1
sqlparse==0.5.3
